+++ next-docs/.env.production
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_ELIXIR_ENABLED=true
+NEXT_PUBLIC_ELIXIR_BASE_URL=http://localhost:4000/
